+++ CHANGES.md
@@ -3,4 +3,5 @@
 ## Ckeditor 4.1.0
   
   * Update ckeditor to 4.4.2 Full Package version (72 Plugins, 24 Jun 2014)
+  * Copy non-digested assets of ckeditor after assets precompile rake task
   * Starting gem changelog